remove flakiness from scale test (#27)
This will reduce flakiness of the system tests. Ocasionally the system
test won't scale to 3 replicas because it's too fast at processing 500
requests.

So instead the deployment is patched to wait for 10 seconds before
starting the main container. This causes the queue to be build up large
enough and cause it to scale up to 3 replicas.