Fix #67 only leader should do scaling (#72)
Only the leader should be responsible for scale up and scale down. The
only exception is "AtLeastOne" call in proxy handler. Any other scale
ups / scale down should not be allowed in the replicas that are not the
leader. That's why UpdateState should not be calling compareScales which
can cause scale ups/downs.