Add integration tests that ensure quickstart works (#167)
Simple e2e test written in bash. This ensures that the steps in the
README won't be broken by other PRs. More advanced e2e tests can be
written in golang. The tests succeed in ~4-10 minutes depending on how
fast the download of gemma 2b is.

In some cases the download may take longer than the timeout.