Recreate KubeAI Pods on change to ConfigMap & add missing ConfigMap field (#214)
Restart KubeAI to pick up changes to configmap.
See:
https://helm.sh/docs/howto/charts_tips_and_tricks/#automatically-roll-deployments

Hot-reload was considered - however it gets complicated b/c of
long-running processes that get started such as messengers and
autoscaling. We can punt on that until later.

Related to #151 which can be implemented seperately.

Also added `.messaging` to ConfigMap (it was omitted).