Multi-turn chat benchmark implemented in Go (#414)
Create a new benchmark in Go to capture metrics from a multi-turn chat
scenario that were not possible with the k6 benchmark (i.e. TTFT).

During the process of benchmarking, a major issue was found the the
`PrefixHash` load balancing strategy: when Pods were terminated
mid-load, the in-flight requests would be double-decremented from the
`totalInFlight` measurement - throwing the "bounded load" check off for
future requests. This issue was fixed in this PR.

---------

Co-authored-by: Sam Stoelinga <sammiestoel@gmail.com>