Support multiple Kubeai installation in different namespace (#518)
At the moment it is not possible to install multiple Kubeai instances in
the same cluster and in different namespaces due to the way helm handles
CRDs.

In this PR I added the _crds.enabled_ value to control the installation
of CRDs.

An alternative would be to move the "crds" folder to the root of the
Chart and use the helm flag --skip-crds to disable its installation.
However, this involves having to manually manage the CRDs update.

---------

Signed-off-by: ffais <ffais@fbk.eu>